feat: integrate MainControlsBar into MapView and enhance TrackEditor with track prop
